% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  account,
  age,
  example_no,
  function_name,
  lambda,
  message,
  mu,
  n_replicates,
  session,
  t,
  verbose
)
}
\arguments{
\item{account}{a peregrine account}

\item{age}{the age of the phylogeny}

\item{example_no}{number of the example}

\item{function_name}{name of the function}

\item{lambda}{speciation rate}

\item{message}{a message to be print}

\item{mu}{extinction rate}

\item{n_replicates}{number of replicates}

\item{session}{a ssh session}

\item{t}{time}

\item{verbose}{choose if you want to print the output or not}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Documentation by Giovanni Laudanno,
}
