% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dd_experiment.R
\name{errors_vs_dd}
\alias{errors_vs_dd}
\title{Execute pirouette's dd experiment}
\usage{
errors_vs_dd(n_replicates = 2)
}
\arguments{
\item{n_replicates}{number of replicates}
}
\value{
nothing
}
\description{
Execute pirouette's dd experiment
}
\author{
Giovanni Laudanno
}
