% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors_vs_ntaxa.R
\name{errors_vs_ntaxa}
\alias{errors_vs_ntaxa}
\title{Errors_vs_ntaxa}
\usage{
errors_vs_ntaxa(n_replicates = 2)
}
\arguments{
\item{n_replicates}{number of replicates}
}
\value{
nothing
}
\description{
Errors_vs_ntaxa
}
\author{
Giovanni Laudanno
}
