% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{open_example_script}
\alias{open_example_script}
\title{run pirouette example}
\usage{
open_example_script(example_no)
}
\arguments{
\item{example_no}{number of the example}
}
\value{
nothing
}
\description{
run pirouette example
}
\author{
Giovanni Laudanno
}
