% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_it}
\alias{run_it}
\title{run pirouette example}
\usage{
run_it(
  project_name = get_pkg_name(),
  function_name = "errors_vs_ntaxa",
  n_replicates = 2,
  account = "p274829",
  session = NA
)
}
\arguments{
\item{function_name}{name of the function}

\item{n_replicates}{number of replicates}

\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
run pirouette example
}
\author{
Giovanni Laudanno
}
