% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{upload_bash_scripts}
\alias{upload_bash_scripts}
\title{Export cluster scripts}
\usage{
upload_bash_scripts(
  project_name = get_pkg_name(),
  account = "p274829",
  session = NA
)
}
\arguments{
\item{account}{a peregrine account}

\item{session}{a ssh session}
}
\value{
nothing
}
\description{
Export cluster scripts
}
\author{
Giovanni Laudanno
}
